+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Views and ViewControllers
 
+- `InnerShadowView` is added
 - `InteractivePoppingNavigationController` is added
 
 ### [3.14.0(45)](https://github.com/VakhoKontridze/VCore/releases/tag/3.14.0) — *2022 07 12*